Splitting api
