+++ temp-server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "temp-server"
-version = "0.1.2"
+version = "0.1.3"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
