+++ temp-server/src/module/config.rs
@@ -1,0 +1,27 @@
+use anyhow::Result;
+use serde::{Deserialize, Serialize};
+use std::{fs::File, io::Read};
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct Config {
+    pub server: ServerConfig,
+}
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct ServerConfig {
+    pub addr: String,
+    pub port: u64,
+    pub url: String,
+}
+
+impl Config {
+    pub fn new(path: &str) -> Result<Self> {
+        let mut file = File::open(path)?;
+
+        let mut str = String::new();
+        file.read_to_string(&mut str)?;
+
+        let config: Config = toml::from_str(&str)?;
+        Ok(config)
+    }
+}
