+++ temp-server/src/module/mod.rs
@@ -1,0 +1,2 @@
+pub mod config;
+pub mod rpc;
